+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `only-delivery` as default shippingSelection value
+
 ## [1.10.0] - 2025-10-28
 
 ### Added